+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - split seq-repo load function into new asset
+- apply all activity filters to all extractors with activities
+- refactor filter functionality to filter module
 
 ### Deprecated